review: act on OngChia's round
- tmx_states, metric_fields: the docstring wordings he suggested, taken as
  proposed.
- compute_weight_factors: drop `_compute_z1_z2_z3`, left behind by the gt4py
  conversion and called from nowhere.
- serialbox: drop the `ta_phy` accessor. Nothing reads it, and the name is a
  trap: in ICON `ta_phy` is a member of the tendency list (mo_aes_phy_memory.f90
  l.544, [K/s]), but the field serialized under that name in 'tmx-entry' is
  bit-identical to `ta`.
- the ISO 8601 duration parser moves to common/utils/time_utils.py, with its
  unit tests; the driver and the tmx fixtures each had a copy.
