+++ model/common/src/icon4py/model/common/metrics/metric_fields.py
@@ -976,18 +976,16 @@ def compute_height_above_ground(
 
     Port of ``compute_geopotential_height_above_ground`` (mo_vdf_atmo.f90):
 
-        ghf(jk) = z_mc(jk) - z_ifc(nlevp1)
+        height_above_ground = z_mc - z_ifc[nlevp1]
 
-    Despite the Fortran name the result is a geometric height in meters;
-    gravity is only applied by the users of the field (e.g.
-    ``compute_dry_static_energy``). Computed with numpy because GT4Py offsets
-    are relative and cannot address the fixed absolute surface row of ``z_ifc``.
+    Computed with numpy because GT4Py offsets are relative and cannot address
+    the fixed absolute surface row of ``z_ifc``.
 
     Args:
         z_mc: geometric height of the full levels [m]
         z_ifc: geometric height of the half levels [m], surface at the last row
 
     Returns:
-        height of the full levels above the surface [m]
+        height_above_ground: height of the full levels above the surface [m]
     """
     return z_mc - z_ifc[:, -1:]
